requirements.txt
asgiref==3.5.2
astroid==2.12.13
autopep8==2.0.0
black==22.10.0
certifi==2022.9.24
click==8.1.3
cloudinary==1.30.0
colorama==0.4.6
dill==0.3.6
Django==4.1.3
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
gunicorn==20.1.0
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
mypy-extensions==0.4.3
mysqlclient==2.1.1
pathspec==0.10.2
platformdirs==2.5.4
pycodestyle==2.10.0
pylint==2.15.7
python-decouple==3.6
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
tzdata==2022.6
urllib3==1.26.13
whitenoise==6.2.0
wrapt==1.14.1
